Move Interaction to its own file
